move UserSessionSelfUserClientDelegate
protocol
